Fix u16 overflow in IPv4PacketBuffer
When the buffer is full, self.head equals 0x10000, so converting its
value to u16 results in 0.

As a consequence, length <= self.head was mistakenly considered false,
so the packet was never consumed.

As a consequence, the next call to read_from() requested a read with
a buffer of 0 bytes, so 0 bytes were read… which was interpreted as EOF,
causing client disconnection.

See <https://github.com/Genymobile/gnirehtet/issues/43>.

+++ relay-rust/src/relay/ipv4_packet_buffer.rs
@@ -42,7 +42,7 @@ impl Ipv4PacketBuffer {
     fn available_packet_length(&self) -> Option<u16> {
         if let Some((version, length)) = ipv4_header::peek_version_length(&self.buf) {
             assert!(version == 4, "Not an Ipv4 packet, version={}", version);
-            if length <= self.head as u16 {
+            if length as usize <= self.head {
                 // full packet available
                 Some(length)
             } else {
